modernize project
- switch from poetry to uv
- modernize README
- modernize docstring content
- switch from tabs to spaces
- switch from numpy-style docstrings to google style
- fully type project and forbid the use of Any
- disable old ruff ignores
- refactor CLI entrypoint into `__main__.py`
