+++ README.md
@@ -22,20 +22,32 @@ Randfacts is a Python module that generates random facts. You can use `randfacts
 
 randfacts can either be installed via pip or via the AUR, whichever way you prefer.
 
-### Installation via pip:
+### Installation via uv/pipx/pip:
 
 ```sh
+# Install with UV:
+$ uv tool install randfacts
+
+# Or just try it out:
+$ uvx randfacts
+
+# Install with pipx:
+$ pipx install randfacts
+
+# Install with bare pip:
 $ pip3 install randfacts
 ```
 
 ### Installation via AUR:
 
 Via your AUR helper, like paru:
+
 ```sh
 $ paru -S python-randfacts
 ```
 
 Or manually
+
 ```sh
 $ git clone https://aur.archlinux.org/python-randfacts.git && cd python-randfacts
 $ makepkg -si
@@ -48,10 +60,12 @@ import randfacts
 x = randfacts.get_fact()
 print(x)
 ```
+
 The above example will print a random fact like:
 `Penguins can't taste sweet or savory flavors, only sour and salty ones`
 
 This package has a filter option to filter out potentially inappropriate facts. The filter is on by default. To disable the filter, you can just set the `filter_enabled` parameter to `False`.
+
 ```python
 from randfacts import get_fact
 print(get_fact(False))
@@ -66,7 +80,6 @@ print(get_fact(only_unsafe=True))
 ```
 
 If you want to access the list of facts directly, you can just import the `safe_facts`, `unsafe_facts`, or `all_facts` lists from the randfacts module.
-
 
 ## Command line usage
