--- randfacts/__main__.py
@@ -1,3 +1,0 @@
-from .randfacts import _cli_entrypoint  # pyright: ignore[reportPrivateUsage]
-
-_cli_entrypoint()
