+++ tests/test_general.py
@@ -8,123 +8,123 @@ import pytest
 
 sys.path.insert(1, str(pathlib.Path(__file__).parents[1]))
 from randfacts import (
-	randfacts,  # local randfacts instead of installed version
+    randfacts,  # local randfacts instead of installed version
 )
 
 
 def test_get_fact() -> None:
-	"""Make sure get_fact works without extra arguments."""
-	assert isinstance(randfacts.get_fact(), str), "get_fact() must return a string"
+    """Make sure get_fact works without extra arguments."""
+    assert isinstance(randfacts.get_fact(), str), "get_fact() must return a string"
 
 
 def test_get_fact_filter_off() -> None:
-	"""Make sure get_fact works with filter off."""
-	assert isinstance(randfacts.get_fact(filter_enabled=False), str), (
-		"get_fact() must return a string"
-	)
+    """Make sure get_fact works with filter off."""
+    assert isinstance(randfacts.get_fact(filter_enabled=False), str), (
+        "get_fact() must return a string"
+    )
 
 
 def test_get_fact_only_unsafe() -> None:
-	"""Make sure get_fact works with only unsafe."""
-	assert isinstance(randfacts.get_fact(only_unsafe=True), str), (
-		"get_fact() must return a string"
-	)
+    """Make sure get_fact works with only unsafe."""
+    assert isinstance(randfacts.get_fact(only_unsafe=True), str), (
+        "get_fact() must return a string"
+    )
 
 
 def test_all_facts_list() -> None:
-	"""Test that all_facts list is present in the module."""
-	assert isinstance(randfacts.all_facts, list), "all_facts must be a list"
+    """Test that all_facts list is present in the module."""
+    assert isinstance(randfacts.all_facts, list), "all_facts must be a list"
 
 
 def test_safe_facts_list() -> None:
-	"""Test that safe_facts list is present in the module."""
-	assert isinstance(randfacts.safe_facts, list), "safe_facts must be a list"
+    """Test that safe_facts list is present in the module."""
+    assert isinstance(randfacts.safe_facts, list), "safe_facts must be a list"
 
 
 def test_unsafe_facts_list() -> None:
-	"""Test that unsafe_facts list is present in the module."""
-	assert isinstance(randfacts.unsafe_facts, list), "unsafe_facts must be a list"
+    """Test that unsafe_facts list is present in the module."""
+    assert isinstance(randfacts.unsafe_facts, list), "unsafe_facts must be a list"
 
 
 def test_cli_no_args() -> None:
-	"""Test that a basic randfacts CLI call will work."""
-	child = subprocess.Popen(["python3", "-m", "randfacts"], stdout=subprocess.DEVNULL)
-	child.communicate()
-	assert child.returncode == 0, "`python3 -m randfacts` must return with exit code 0"
+    """Test that a basic randfacts CLI call will work."""
+    child = subprocess.Popen(["python3", "-m", "randfacts"], stdout=subprocess.DEVNULL)
+    _ = child.communicate()
+    assert child.returncode == 0, "`python3 -m randfacts` must return with exit code 0"
 
 
 def test_cli_script_installed() -> None:
-	"""Test that the `randfacts` script is installed to the PATH."""
-	child = subprocess.Popen(["randfacts"], stdout=subprocess.DEVNULL)
-	child.communicate()
-	assert child.returncode == 0, "`randfacts` must return with exit code 0"
+    """Test that the `randfacts` script is installed to the PATH."""
+    child = subprocess.Popen(["randfacts"], stdout=subprocess.DEVNULL)
+    _ = child.communicate()
+    assert child.returncode == 0, "`randfacts` must return with exit code 0"
 
 
 def test_cli_unsafe_args() -> None:
-	"""Test that CLI with --unsafe works."""
-	child = subprocess.Popen(
-		["python3", "-m", "randfacts", "--unsafe"],
-		stdout=subprocess.DEVNULL,
-	)
-	child.communicate()
-	assert child.returncode == 0, (
-		"`python3 -m randfacts --unsafe` must return with exit code 0"
-	)
+    """Test that CLI with --unsafe works."""
+    child = subprocess.Popen(
+        ["python3", "-m", "randfacts", "--unsafe"],
+        stdout=subprocess.DEVNULL,
+    )
+    _ = child.communicate()
+    assert child.returncode == 0, (
+        "`python3 -m randfacts --unsafe` must return with exit code 0"
+    )
 
 
 def test_cli_mixed_args() -> None:
-	"""Test that CLI with --mixed works."""
-	child = subprocess.Popen(
-		["python3", "-m", "randfacts", "--mixed"],
-		stdout=subprocess.DEVNULL,
-	)
-	child.communicate()
-	assert child.returncode == 0, (
-		"`python3 -m randfacts --mixed` must return with exit code 0"
-	)
+    """Test that CLI with --mixed works."""
+    child = subprocess.Popen(
+        ["python3", "-m", "randfacts", "--mixed"],
+        stdout=subprocess.DEVNULL,
+    )
+    _ = child.communicate()
+    assert child.returncode == 0, (
+        "`python3 -m randfacts --mixed` must return with exit code 0"
+    )
 
 
 def test_cli_version() -> None:
-	"""Test that CLI with --version returns the correct version."""
-	child = subprocess.run(
-		["python3", "-m", "randfacts", "--version"],
-		stdout=subprocess.PIPE,
-		universal_newlines=True,  # noqa: UP021 for python 3.6 support
-		check=True,
-	)
+    """Test that CLI with --version returns the correct version."""
+    child = subprocess.run(
+        ["python3", "-m", "randfacts", "--version"],
+        stdout=subprocess.PIPE,
+        universal_newlines=True,  # noqa: UP021 for python 3.6 support
+        check=True,
+    )
 
-	assert child.stdout.strip() == randfacts.__version__, (
-		f"`python3 -m randfacts --version` must return {randfacts.__version__}"
-	)
+    assert child.stdout.strip() == randfacts.__version__, (
+        f"`python3 -m randfacts --version` must return {randfacts.__version__}"
+    )
 
 
 def test_main_entrypoint() -> None:
-	"""Test the main entrypoint in randfacts.py."""
-	# Path to the module or script you want to test
-	script_path = (
-		pathlib.Path(__file__).resolve().parents[1] / "randfacts" / "randfacts.py"
-	)
+    """Test the main entrypoint in randfacts.py."""
+    # Path to the module or script you want to test
+    script_path = (
+        pathlib.Path(__file__).resolve().parents[1] / "randfacts" / "randfacts.py"
+    )
 
-	# Run the script as a subprocess
-	result = subprocess.run(  # noqa: UP022 for python 3.6
-		["python", str(script_path)],
-		stdout=subprocess.PIPE,
-		stderr=subprocess.PIPE,
-		universal_newlines=True,  # noqa: UP021 for python 3.6 support
-		check=False,
-	)
+    # Run the script as a subprocess
+    result = subprocess.run(  # noqa: UP022 for python 3.6
+        ["python", str(script_path)],
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+        universal_newlines=True,  # noqa: UP021 for python 3.6 support
+        check=False,
+    )
 
-	# Assert the subprocess exits successfully
-	assert result.returncode == 0, f"Script failed with stderr: {result.stderr}"
+    # Assert the subprocess exits successfully
+    assert result.returncode == 0, f"Script failed with stderr: {result.stderr}"
 
 
 @pytest.mark.parametrize("bad_char", ["‘", "’", "“", "”", "…", "—"])  # noqa: RUF001
 def test_invalid_characters(bad_char: str) -> None:
-	"""Make sure no invalid characters are present in the fact lists.
+    """Make sure no invalid characters are present in the fact lists.
 
-	If this test fails, try running `fix_encoding.py`
-	"""
-	for index, fact in enumerate(randfacts.all_facts):
-		assert bad_char not in fact, (
-			f"Bad character '{bad_char}' found in fact at index {index}"
-		)
+    If this test fails, try running `fix_encoding.py`
+    """
+    for index, fact in enumerate(randfacts.all_facts):
+        assert bad_char not in fact, (
+            f"Bad character '{bad_char}' found in fact at index {index}"
+        )
